requirements.txt
jpegio==0.2.8
numpy==2.4.4
Pillow==12.2.0
PySide6==6.10.2
pytest==9.0.2
pytest-qt==4.5.0
cryptography==46.0.7